requirements.txt
click==8.1.3
colorama==0.4.4
Flask==2.1.2
Flask-MySQL==1.5.2
Flask-MySQLdb==1.0.1
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql==0.0.3
mysqlclient==2.1.0
PyMySQL==1.0.2
Werkzeug==2.1.2
zipp==3.8.0
